refactor: redid the file structure
