+++ src/pages/index.js
@@ -1,4 +1,4 @@
 
-export { default as Home } from './Home'
-export { default as About } from './About'
-export { default as Contact } from './Contact'
+export { default as Home } from './home/Home'
+export { default as About } from './about/About'
+export { default as Contact } from './contact/Contact'
